rename penevent module
