+++ crates/rnote-compose/src/builders/fociellipsebuilder.rs
@@ -3,7 +3,7 @@ use super::shapebuildable::{ShapeBuilderCreator, ShapeBuilderProgress};
 use super::ShapeBuildable;
 use crate::constraints::ConstraintRatio;
 use crate::ext::AabbExt;
-use crate::penevents::{PenEvent, PenState};
+use crate::penevent::{PenEvent, PenState};
 use crate::penpath::Element;
 use crate::shapes::Ellipse;
 use crate::style::{indicators, Composer};
