+++ crates/rnote-compose/src/builders/penpathmodeledbuilder.rs
@@ -1,9 +1,9 @@
 // Imports
 use super::{PenPathBuildable, PenPathBuilderCreator, PenPathBuilderProgress};
-use crate::penevents::PenEvent;
 use crate::penpath::{Element, Segment};
 use crate::style::Composer;
 use crate::Constraints;
+use crate::PenEvent;
 use crate::{PenPath, Style};
 use ink_stroke_modeler_rs::{
     ModelerInput, ModelerInputEventType, ModelerParams, PredictionParams, StrokeModeler,
