+++ crates/rnote-compose/src/builders/quadrantcoordsystem2dbuilder.rs
@@ -4,7 +4,7 @@ use std::time::Instant;
 
 use super::shapebuildable::{ShapeBuilderCreator, ShapeBuilderProgress};
 use super::ShapeBuildable;
-use crate::penevents::{PenEvent, PenState};
+use crate::penevent::{PenEvent, PenState};
 use crate::penpath::Element;
 use crate::shapes::Line;
 use crate::style::{indicators, Composer};
