+++ crates/rnote-compose/src/lib.rs
@@ -14,7 +14,7 @@ pub mod constraints;
 /// module for extension traits for foreign types
 pub mod ext;
 /// module for pen events
-pub mod penevents;
+pub mod penevent;
 /// module for pen paths
 pub mod penpath;
 /// utilities for serializing / deserializing
@@ -33,6 +33,7 @@ pub mod utils;
 // Re-exports
 pub use color::Color;
 pub use constraints::Constraints;
+pub use penevent::PenEvent;
 pub use penpath::PenPath;
 pub use shapes::Shape;
 pub use splitorder::SplitOrder;
