+++ crates/rnote-engine/src/engine/mod.rs
@@ -25,7 +25,7 @@ use futures::channel::{mpsc, oneshot};
 use gtk4::gsk;
 use p2d::bounding_volume::{Aabb, BoundingVolume};
 use rnote_compose::ext::AabbExt;
-use rnote_compose::penevents::{EventPropagation, PenEvent, ShortcutKey};
+use rnote_compose::penevent::{EventPropagation, PenEvent, ShortcutKey};
 use rnote_compose::shapes::Shapeable;
 use rnote_compose::SplitOrder;
 use serde::{Deserialize, Serialize};
