+++ crates/rnote-engine/src/pens/shortcuts.rs
@@ -1,6 +1,6 @@
 // Imports
 use super::PenStyle;
-use rnote_compose::penevents::ShortcutKey;
+use rnote_compose::penevent::ShortcutKey;
 use serde::{Deserialize, Serialize};
 use std::collections::HashMap;
 use std::ops::{Deref, DerefMut};
