+++ crates/rnote-ui/src/canvas/input.rs
@@ -1,9 +1,9 @@
 // Imports
 use super::RnCanvas;
 use gtk4::{gdk, glib, prelude::*, Native};
-use rnote_compose::penevents::ShortcutKey;
-use rnote_compose::penevents::{KeyboardKey, PenState};
-use rnote_compose::penevents::{ModifierKey, PenEvent};
+use rnote_compose::penevent::ShortcutKey;
+use rnote_compose::penevent::{KeyboardKey, PenState};
+use rnote_compose::penevent::{ModifierKey, PenEvent};
 use rnote_compose::penpath::Element;
 use rnote_engine::ext::EventPropagationExt;
 use rnote_engine::pens::penholder::BacklogPolicy;
