+++ crates/rnote-ui/src/settingspanel/mod.rs
@@ -15,7 +15,7 @@ use gtk4::{
     Widget,
 };
 use num_traits::ToPrimitive;
-use rnote_compose::penevents::ShortcutKey;
+use rnote_compose::penevent::ShortcutKey;
 use rnote_engine::document::background::PatternStyle;
 use rnote_engine::document::format::{self, Format, PredefinedFormat};
 use rnote_engine::ext::GdkRGBAExt;
